XW-4310 | fixes test for trending-videos-item
